@@ -1,8 +1,9 @@
+// @ts-nocheck
 import React, { useState, useCallback, useMemo, useEffect, FC } from "react";
-import { Editable, withReact, useSlate, Slate } from "slate-react";
-import { Editor, Transforms, createEditor, Node } from "slate";
+import { Editable, withReact, useSlate, Slate, ReactEditor } from "slate-react";
+import { Editor, Transforms, createEditor, Node, BaseEditor } from "slate";
 import isHotkey from "is-hotkey";
-import { withHistory } from "slate-history";
+import { withHistory, HistoryEditor } from "slate-history";
 import { Button, Toolbar } from "./RichTextControls";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
@@ -21,6 +22,18 @@ export const getTextFromNodes = (nodes: Node[]) => {
   return nodes.map((n: Node) => Node.string(n)).join("\n");
 };
 
+// https://docs.slatejs.org/concepts/12-typescript#defining-editor-element-and-text-types
+type CustomElement = { type: "paragraph"; children: CustomText[] };
+type CustomText = { text: string; bold?: true };
+
+declare module "slate" {
+  interface CustomTypes {
+    Editor: BaseEditor & ReactEditor & HistoryEditor;
+    Element: CustomElement;
+    Text: CustomText;
+  }
+}
+
 const HOTKEYS: { [keyName: string]: string } = {
   "mod+b": "bold",
   "mod+i": "italic",
@@ -30,9 +43,9 @@ const HOTKEYS: { [keyName: string]: string } = {
 const initialValue = [
   {
     type: "paragraph",
-    children: [{ text: "" }],
+    children: [{ text: "Enter your post here." }],
   },
-];
+] as CustomElement[];
 const LIST_TYPES = ["numbered-list", "bulleted-list"];
 
 class RichEditorProps {
@@ -47,17 +60,20 @@ const RichEditor: FC<RichEditorProps> = ({
   sendOutBody,
 }) => {
   const [value, setValue] = useState<Node[]>(initialValue);
-  const renderElement = useCallback((props) => <Element {...props} />, []);
-  const renderLeaf = useCallback((props) => <Leaf {...props} />, []);
+  const renderElement = useCallback((props: any) => <Element {...props} />, []);
+  const renderLeaf = useCallback((props: any) => <Leaf {...props} />, []);
   const editor = useMemo(() => withHistory(withReact(createEditor())), []);
 
   useEffect(() => {
-    console.log("existingBody", existingBody);
     if (existingBody) {
-      setValue(JSON.parse(existingBody));
+      setValue([
+        {
+          type: "paragraph",
+          text: existingBody,
+        },
+      ]);
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [existingBody]);
+  }, []);
 
   const onChangeEditorValue = (val: Node[]) => {
     setValue(val);
@@ -82,7 +98,7 @@ const RichEditor: FC<RichEditorProps> = ({
         className="editor"
         renderElement={renderElement}
         renderLeaf={renderLeaf}
-        placeholder="Enter your post here."
+        placeholder="Enter some rich text…"
         spellCheck
         autoFocus
         onKeyDown={(event) => {
